feat: add api versioning (API-4) (#3)
Bootstraps API versioning. All controllers have been split into subfolders that correspond to the version of they API they govern. Unit test will follow the same pattern when covering these resources. In the event no API version is specified in the URL, the application will use the most recent stable version of the API to handle the request.

The current version of the API will start at `v0` to indicate that it is still under development. When the API is in a stable state and the majority of the development finished, this will be bumped to `v1`. An endpoint to check the server's status has also been implemented as part of the bootstrapping for API versioning. This endpoint is set to be handled at the root of the API for both `v0` (`/api/v0`), as well as the unversioned root domain (`/api/`).