ci: integrate GitHub Actions as CI platform
Four workflows have been introduced here;

* build-branch
* build-release
* create-release
* pull-request

The build-branch and build-release workflows will simply run the application's unit tests, and then build the docker container in the event the test suite has successfully passed. While these two workflows ultimately accomplish the same task, they are different in the aspect that building a branch will use the first 7 characters of the commit hash as the Docker tag, while the release will use the git tag version as the Docker tag. The build-branch workflow will also have the Docker tag prefixed with the branch name. For example, a branch's docker container would be named something along the lines of `viiidb-api:develop-2d70624`, while a release's Docker container would be named something like `viiidb-api:1.0.0`.

The only reason for splitting these two workflows is to keep things a bit more self-contained. If these workflows are grouped together, conditionals would need to be put in place to manage which workflow runs on which git ref. Having conditionals in workflows should be kept to a minimum as whenever they are "skipped" they appear as a "passing" test on the CI flag next to the commit. This type of functionality is undesirable.

In addition to the two workflows mentioned above, there is also a workflow for create-release, as well as pull-request. Touching on pull-request first, this workflow will simply run unit tests whenever a pull request is opened. Nothing special or fancy there. The create-release workflow, however, will take care of issuing an official release when a pull request is merged into the `master` branch. Creating this release will invoke the standard-version NPM package to automate the process of versioning, as well as changelog generation. Upon successfully completing this, standard-version will commit the changes back to the `master` branch and proceed to create a git tag with the current version of the application. This tag will also be pushed to source control where it will be picked up by the Publish Release workflow job, linking the newly created tag to an official release.

It is important to note that the version that is released may be overridden by passing the `[release as]` flag to either the commit title, or body. For example, if it was desirable to create a release for `v0.1.0` the flag `[release as 0.1.0]` would need to be included in the commit (notice the omission of the `v` character).

To support the automated release cycle using standard-version, a `.versionrc` file has been included which store the configuration for standard-version. Within this configuration there is a `sed` script that will run once the changelog has been generated to standardize the changelog headings so they may be picked up by the release job properly.