feat: add seed rank resource (API-8) (#6)
* feat: add seed rank resource

[Seed Rank Resources]

Seed Rank resource have been added!

To support this addition, the application has been abstracted out a bit from the traditional "Laravel" patterns.

Services are being used as a proxy to fetch data from the database, rather than using Models directly through a Controller. Using Services will allow the system to be flexible in the event it is ever extended to use a different data layer, similar to a repository pattern. However, alongside this, it will also allow the system to reuse the same chunks of code through the same class API. Instead of having to repeat all of the queries on every controller, the system can reference the same method across multiple Controllers. This is handy for versioning the API.

Transformers have also been added as a way to standardize the JSON objects being sent to the client for consumption. These resources are versioned and will correspond to the Controller and Endpoint of the same version. The idea here is that the Services will allow the Controllers to keep the same data layer API across versions, while the Controllers will use Transformers to standardize the output to the client. Transformers are used here instead of JSON Resources (Laravel built-in) simply due to the fact that Laravel's JSON Resources are tightly coupled to Models. Using a Transformer allows the API to be agnostic to the data layer when standardizing the client output.

[Exception Handler Updates]

The Exception Handler has been updated to do a couple of things. Since the application will be purely an API, it has been modified to append the `accept: application/json` header to every exception thrown from any `/api/*` endpoint. From there, it has also been modified to send a custom formatted JSON response in place of the exception. This JSON response is constructed in the same way that Laravel's native methods create JOSN exceptions, albeit it has been abstracted to a sub-level `errors` key, while basic request metadata has been placed at the root of the response.

[Development Docker Configuration]
The `docker-compose.yaml` file has been updated to use the `VIRTUAL_HOST` environment variable directly instead of through the ENV. This will allow for the testing environment to be constructed properly.

For context; if the environment file is injected directly, the `.env.testing` configuration will not be loaded during testing. Instead, the `.env` file will be loaded in as a top level configuration, preventing any testing configuration overrides. The only value needed for development is the `VIRTUAL_HOST` value, so it is the only value included now.

[PHP-CS-Fixer]

The `.php-cs-fixer.php` file has been updated to include the following fixers;

* `not_operator_with_succesor_space`
    * This is configured so that a single space is applied after the logical not (`!`) operator.
* `binary_operator_spaces`
    * This is actually broken for aligning arrays. With this being the case, the array formatting has been explicitly disabled so that arrays may be aligned vertically manually.
* `yoda_style`
    * This is included as part of a preset. However, it has been explicitly disabled so that conditionals can be written in a way that allows for fluent readability.

[Misc]

* The `.env.example` file has been updated to include the default development configuration for the database.
* The dependency `webpatser/laravel-uuid` was included so that Models may use UUIDs over integer based index's. Laravel _does_ have a UUID library that it ships with. However, I prefer this one.
* The `.gitignore` has been updated to exclude test reports from source code.

* build: update build process to use `-dev` container images

The development environment was updated to remove the `dev.Dockerfile` from the root of the project and instead use a container specifically made for development. Alongside this, the GitHub Actions configuration was updated to follow suit, using the same Docker image to run unit tests for the application.