refactor!: remove `/api` prefix from endpoints (#9)
It doesn't make much sense to have `/api` as a prefix to the endpoints as the entire application itself is strictly an API. Referencing `api.viiidb.com/api` just feels redundant. This removes the `/api` prefix from the endpoints and restructures the codebase to follow suit by also removing `api` subfolders.

BREAKING CHANGE: The `/api` prefix for endpoints has been removed due to redundancy. Endpoints with the `/api` prefix will not longer work.