chore: trigger ci build process (#14)
A previous commit was used to resolve an issue where releases were not triggering the build process. The flag used to skip the ci workflow was used in the commit message as an example. Due to this, the ci workflow was skipped when it actually should have been run. This commit simply triggers the ci workflow so that the container may be built and released.