refactor: explicitly define api routes
Instead of using resource routes and only plucking the ones needed, routes are now specified explicitly. Doing so allows route names to be excluded without the need to strip them from each individual endpoint in a resource route. Excluding the route names will alow the application to serialize the routes when the container is created. It also doesn't make much sense to use a resource route if only two of the endpoints in a resource route are needed.