feat: allow records to be filtered by the value of a given field
A new trait has been added that will allow models to filter records based on the values of the fields that have been marked as "filterable". Users may pass a `key=value` pair to the query string representing the data they would like to retrieve from the API. For example; https://api.viiidb.com/v0/seed-ranks?rank=1 will return all of the SeeD Ranks where the `rank` column is equal to `1`. Likewise, https://api.viiidb.com/v0/seed-ranks?salary=500 will return all of the SeeD Ranks where the `salary` column is equal to `500`.

In the event a user does not know the exact value of the record they would like to retrieve, the `like:` prefix may be included as part of the `value` in order to return all records that are similar to the value specified. For example; https://api.viiidb.com/v0/seed-ranks?rank=like:1 will return all of the SeeD Ranks where the `rank` column is similar to `1` (1, 10, 12, 21, etc...). Likewise, https://api.viiidb.com/v0/seed-ranks?salary=like:500 will return all of the SeeD Ranks where the `salary` column is similar to `500` (500, 1500, 2500, 15000, etc...).

This can be set up by simply including the `App\Traits\FiltersRecordsByFields` trait on the model and instructing it to use said trait. From there, an array of fields should be defined via the `filterableFields` trait on the model. The associated service will be responsible for filtering these records for the end user.