feat: add seed test resource (API-10) (#20)
The SeeD Test resource has been added!

This resource includes the same scaffolding and functionality as with the SeeD Rank resource, and is available via the `/seed-tests` endpoint. Note that this is only the SeeD Test resource itself, and does not include the associated Test Questions. The Test Questions would "technically" be considered their own resource and will be included as part of a different PR. When the Test Questions resource is added, they will also be tied to the SeeD Test resource as part of the same PR.