refactor: extract common logic to base model (API-79) (#23)
* refactor: extract common logic to base model

This will introduce a base `App\Models\Model` class for all other models to extend. This base class includes common functionality and configurations used throughout each model in an attempt to keep the codebase relatively clean and stable. As an addition, various traits have been updated to include their local properties within the trait itself. This was not possible previously as PHP would not allow traits to be defined on both the trait, as well as the class which implements said trait. Properties of traits may only be defined on classes that _inherit_ another class which uses a trait.

The base model includes the following traits out of the box;

* `App\Traits\FiltersRecordsByFields`
* `App\Traits\LoadsRelationsThroughServices`
* `App\Traits\OrdersQueryResults`
* `App\Traits\Searchable`
* `App\Traits\Uuids`

Tests have been updated to follow suit and code coverage remains at 100%.

* fix: use correct type within docblocks for the 'orderBy' properties