feat: introduce exception monitoring via sentry (API-61) (#26)
This introduces exception monitoring via Sentry. When the `SENTRY_LARAVEL_DSN` is configured, all exceptions will be forwarded to Sentry so that they may be captured externally. This will help with alerting in the event something goes wrong with the application. These exceptions may be configured to forward to a specific environment explicitly, otherwise they'll default to the environment that Laravel is running under. Performance monitoring may also be enabled via the use of the `SENTRY_TRACES_SAMPLE_RATE` environment variable.