feat: add inbound request analytics (API-53) (#27)
* feat: add inbound request analytics

This introduces the ability to capture inbound requests which will be used for analytics throughout the application's lifetime. With this being the only use case, I've left the implementation fairly barebones and have excluded the scaffolding for models and such. Instead, we're simply creating a database table and dropping the request data directly into the table as part of the `CaptureInboundRequest` middleware.

The `CaptureInboundRequest` middleware is mapped to the API route group, and will activate _after_ the throttle middleware has processed the request. This will unfortunately skew response time reporting slightly (~5-10ms during testing), but will also result in better stability in the long run as throttled requests won't be clogging up the system.