fix: remove throttle middleware from status endpoint (API-80) (#29)
Kubernetes will probe the application container periodically to ensure that the container is stable. If the probe returns a non-successful status code, Kubernetes will kill the container and spin up a new one. It was discovered that Kubernetes had been periodically restarting the VIIIDB-API container due to throttling limits on the status endpoint. This will simply remove the throttle limitations on the status endpoint alone so that the probes will not restart the container under false conditions.