feat: add status effect resource (API-12) (#33)
This introduces the Status Effect resource. The scaffolding follows suit with other resources within the system, albeit this time I've opted to pass on the overly verbose unit testing. Instead, test for this resource consist of feature tests for the new endpoints that have been added, as well as unit tests for the model and transformer to ensure that the structures in place are correct. Cover coverage still sits at 100% across the entire suite, so this should speed up development time for features moving forward.

The existing models have also been modified to remove the unnecessary Searchable import. For some reason PHPCS didn't catch this and left it in the class.