feat: add stat resource
This introduces the Stat resource. The included scaffolding is a bit more barebones than what is present with typical resources. The Stat resource only includes a model and a transformer. It does not include endpoints or controllers as they will not be used. Instead, this resource will only be used through relationships as a way to standardize commonly used attributes.