feat: add element resource (#39)
Introduces the Element resource to the system. Not much to outline for this as it's all really the same scaffolding as with previous additions. This resource includes two endpoints for the `index` and `show` actions. The status endpoint has also been updated to include the new resource information. Testing isn't as robust as when this project was initially started. Only the necessary test cases have been included to keep development times down and to keep from duplicating test cases when they are not needed. Code coverage is still at 100%, so dropping the duplicate scenarios didn't impact anything on that front.