docs: update changelog to include missing ticket number (API-87) (#43)
When merging the changes for the Elements resource, the tag was unintentionally left out of the commit message. This will simply include the missing ticket number on the changelog that was generated from said commit message to keep a more accurate log. The order of the features for the associated release was also modified so that it reflects the order the changes were made.