feat: introduce meilisearch service via laravel scout
This commit introduces a search service via Laravel Scout. The driver chosen is Meilisearch due to the self-hosting capabilities that should play nicely with Kubernetes.

The included search service will offload the functionality that was being previously developed with raw SQL and provide a fast, easy to use interface for fetching records from the system. Following this update, the existing scaffolding for searching records via SQL should be removed as it is no longer necessary.

Getting up and running with Meilisearch is super easy as well. For development, a Meilisearch Docker container has been added to the global resources in the readme. Simply add this to the `docker-compose` manifest and spin up the resource. For production, a self-hosted instance of Meilisearch should be set up within the Kubernetes cluster so that the records can be indexed properly.

Once the Meilisearch instance has been set up, a console command has been included to index all of the resources that are currently utilizing the search functionality. This can be run with `php artisan search:seed`. The command will seed all of the indexes with data, and will also set up said indexes to use the proper columns and sorting based on what has been configured on each model.