refactor: extract requests away from services
Services previously relied on an instance of `Illuminate\Http\Request` as a function parameter. This does not allow for a very extensible interface and thus the functionality has been abstracted. Services now rely only on the primitive data that the need to fetch data. In this case it was only the requested includes.

These includes have been offloaded to the controller where they will be extracted from the request before getting passed to the service. To take this a step further, a middleware has also been created that will sanitize and convert the values to an array before it even gets to the controller, ensuring the controller will always have an array to work with.

Alongside these changes, some of the unit tests that were not necessary have ben removed.