refactor: general code cleanup and misc refactors (#52)
* Removed unused `Illuminate\Http\Request` imports from controllers
* Removed broadcast and console routes
* Removed unit tests that were no longer necessary
  * Controllers
  * Services
* Refactored the way includes are being processed on the various resources

Regarding the refactor; 

Services previously relied on an instance of `Illuminate\Http\Request` to be passed as a function parameter. This does not allow for a very extensible interface and thus the functionality has now been abstracted. Services now rely only on the primitive data that is needed to fetch data. In this case it was only the requested includes.

These includes have been offloaded to the controller where they will be extracted from the request before getting passed to the service. To take this a step further, a middleware has also been created that will sanitize and convert the values to an array before it even gets to the controller, ensuring the controller will always have an array to work with.