chore: update docker manifest syntax and include refactors with versioning (#56)
* chore: update docker-compose manifest with proper api schema

Updates the `docker-compose.yaml` manifest with an API schema that matches the `3.5` spec. This only impacts the `networks` section of the manifest to specify an external network properly.

* build: include refactor type with .versionrc

* docs: update environment value for meilisearch container

The meilisearch container started throwing errors regarding the value present in the `MEILI_NO_ANALYTICS` environment variable with the newest Docker version. This update ensures that the documentation references the proper format that Docker expects now.