feat: add items resource (#60)
This adds the Items resource to the API. By association, it also adds them to the Meilisearch index. When searching the system for resources, the Items resource will currently be displayed as the top resource among search results. This is due to the frequency of which users will likely need to search for Items within the system.