feat: add location resource (#62)
Pretty large addition this time. With the Locations resource comes the building blocks for future additions. Draw Points, Enemies, Items, Cards, Guardian Forces, etc... All of these will rely upon the Location resource, so extra time was taken to make sure that the structure will work with the coming additions to the system.

The Locations resource utilizes a same-table relationship on itself to handle child/parent areas. That is to say, a Location can both belong to another area, and also have child areas that belong to it. For example; there are 11 different regions throughout the game. Each of these regions contain a subset of overworld areas, of which also contain various other areas such as towns. While they are all still considered locations in the game, they have a set hierarchy in place to help identify where they exist within the game world. This was taken into consideration when designing the relationship structure for the resource.

To get the different relationships between these records, the `region`, `parent` and `locations` methods may be used to return what you would expect them to return. The `region` method will return the current region that a record belongs to. All records will have a region tied to them, whether or not they are nested within the hierarchy. They *are* still associated with a specific region after all regardless of whether the area is in a town or a dungeon. The `parent` method will return the area directly above the current record. If the area is a hotel in a town, then it would be the town that is returned. The `locations` method will return the direct descendants associated with the current record. If the current record in question is a town, it will return the locations present within said town.

Due to this relationship hierarchy and the way that the game was designed, it was necessary to introduce "placeholder" or "container" locations to group some of these locations together. This is an area that doesn't technically exist within the game, but is understood to be within the game world. For example, Balamb Town doesn't have an area named "Balamb Town". Instead, the main area of the town is listed as "Balamb - Town Square". It made more sense to create a "container" record with the name "Balamb Town" and then tie the areas related to Balamb Town to this record instead of "Balamb - Town Square". These records are outlined within the notes section of each applicable record.

All of these records have also been added to the search provide as well. The Locations resource has been placed second in the priority list as this is anticipated to be commonly searched for.