chore(deps): update dependencies for meilisearch and sentry (#64)
Updates Meilisearch from `0.23.1` to `1.0` marking the first "stable" release of the search provider. Also updates Sentry from `2.12` to `3.2`.

The documentation has also been adjusted to reflect the updated dependencies, as well as a change in the default domain name used for local development. This has been moved from `api.local.viiidb.com` to `api.viiidb.local`. Instructions on how to generate an SSL cert for the domain have been updated to follow suit.