ci: include phpstan within ci pipeline (#66)
* ci: include phpstan within ci pipeline

This simply includes PHPStan as part of the CI workflows. Nothing fancy to outline -- PHPStan will simply run after the unit tests when creating pull requests, and when building a branch/release.

* refactor: remove dev container memory limit increase

The memory limit for devleopment containers was being set to 1G within `dev.Dockerfile`. This has since been moved to the base development container and no longer needs to exist within the project.