feat: add rate limiting to api calls (API-82) (#69)
* feat: add rate limiting to api calls

This adds rate limiting support that will apply within a multi-container Kubernetes setup. Using Redis, we can store a central state between all of the deployed containers that will keep track of the IP address that a request comes from, along with the number of requests that said IP has sent over the course of the last 60 seconds. By default, this will allow 60 requests per minute for each IP.

* test: exclude endpoint throttling during testing